Interim changes; I'm starting to solidify things
Finished out the YamlTagCollection and YamlHandleCollection classes.  The
IDictionary implementation in each are explicitly read-only.

Started work on schemas.  Really just have the beginnings of the Failsafe
schema.  A schema (IYamlSchema) will need to be injected into the parser.
(Now that I think about it, my current implementation is wrong for that
to work correctly.  I'll rethink it later, check it in for now.)

Discovered, on re-reading the Yaml 1.2 specification again, that YamlMapping
should be a dictionary of <IYamlNode, IYamlNode>, instead of
<YamlScalar, IYamlNode>.  This is because the specification says that keys are
not restricted to being scalar.  I think this requirement will muck up
YamlMapping a bit, but I have to think about it.

Made the YamlScalar base class YamlScalar<T>, for ease of implementing the
type-specific classes.  I don't know if it will work out that way in the
long run, though.